backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
chromadb==0.5.23
pymupdf==1.25.1
aiosqlite==0.20.0
pydantic-settings==2.7.0
sse-starlette==2.2.1
python-multipart==0.0.20
httpx==0.28.1
